+++ src/Components/common/PelcroNativeButtons/index.js
@@ -1,6 +1,8 @@
-import { userHasAddress } from "../../utils/utils";
+import { saveToMetadataButton } from "./saveToMetadata";
+import { userHasAddress } from "../../../utils/utils";
 
 export const init = (app) => {
+  saveToMetadataButton.init(app);
   const pelcroLoginButtonsByClass = document.getElementsByClassName(
     "pelcro-login-button"
   );
@@ -71,59 +73,6 @@ export const init = (app) => {
         pelcroSubscribeButtonsByClass[j].addEventListener(
           "click",
           app.displaySelectView
-        );
-      }
-    }
-  }
-
-  const pelcroSaveButtonsByClass = document.getElementsByClassName(
-    "pelcro-save-button"
-  );
-
-  const saveToMetadataByButton = (e) => {
-    const { key } = e.currentTarget.dataset;
-    const value = JSON.parse(JSON.stringify(e.currentTarget.dataset));
-    delete value.key;
-    let newVal = "";
-    const pelcroUser = window.Pelcro.user.read();
-    if (
-      pelcroUser.metadata &&
-      pelcroUser.metadata[`metadata_${key}`]
-    ) {
-      const oldValue = pelcroUser.metadata[`metadata_${key}`];
-      if (typeof oldValue === "object" && oldValue.length) {
-        newVal = [...oldValue, value];
-      } else {
-        newVal = [oldValue, value];
-      }
-    } else {
-      newVal = [value];
-    }
-
-    window.Pelcro.user.saveToMetaData(
-      {
-        key,
-        value: newVal,
-        auth_token: window.Pelcro.user.read().auth_token
-      },
-      (err, resp) => {
-        console.log(
-          "window.Pelcro.user.saveToMetaData -> resp",
-          resp
-        );
-      }
-    );
-  };
-
-  if (pelcroSaveButtonsByClass.length !== 0) {
-    for (let j = 0; j < pelcroSaveButtonsByClass.length; j++) {
-      if (
-        pelcroSaveButtonsByClass[j].dataset &&
-        "key" in pelcroSaveButtonsByClass[j].dataset
-      ) {
-        pelcroSaveButtonsByClass[j].addEventListener(
-          "click",
-          saveToMetadataByButton
         );
       }
     }
@@ -260,6 +209,7 @@ export const init = (app) => {
 };
 
 export const authenticatedButtons = (id) => {
+  saveToMetadataButton.authenticated();
   const pelcroLoginByClass = document.getElementsByClassName(
     "pelcro-login-button"
   );
@@ -272,6 +222,8 @@ export const authenticatedButtons = (id) => {
 };
 
 export const unauthenticatedButtons = (id) => {
+  saveToMetadataButton.unauthenticated();
+
   const pelcroLoginByClass = document.getElementsByClassName(
     "pelcro-login-button"
   );