+++ changelog.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog][keep a changelog] and this project adh
 
 ### Added
 
+- Pelcro save and follow button to support loading/active states via classes.
 - Add ecommerce orders section in the user's dashboard.
 - Authorship logo in modals footers now links to Pelcro's homepage.